requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.5
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.1
Flask==3.1.0
idna==3.10
imap-tools==1.8.0
ipython==8.12.3
iso8601==0.1.12
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.0
mysql-connector-python==9.1.0
names==0.3.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
packaging==24.2
pipreqs==0.4.13
platformdirs==4.3.6
ptyprocess==0.7.0
pychallonge==1.11.5
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2019.3
pyzmq==26.2.0
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzlocal==2.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
